pyproject.toml
[tool.poetry]
name = "qlikjwt"
version = "0.1.0"
description = "Create JWT tokens from a private key and a payload"
authors = ["David Wales <david.wales@swsphn.com.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8 <4.0"
typer = "^0.9.0"
PyJWT = "^2.7.0"
cryptography = "^41.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[tool.poetry.scripts]
qlikjwt = 'qlikjwt.qlikjwt:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
